feat(playback): production playback hardening — ticketed proxy, media state machine, reconciliation, telemetry
Completes the in-flight P0 playback work and closes the audit gaps:

- GET /api/v1/tracks/{id}/playback issues short-lived HMAC tickets; stream
  proxy validates them, preserves 200/206/416 + HEAD semantics, and always
  sends private, no-store. Enforcement stays behind
  PlaybackMedia__ReadinessEnforcementEnabled (default off).
- TrackMedia state machine: publish-time validate-and-promote via new
  IMediaReadinessService (fixes uploads being permanently unpublishable);
  transient validation-dependency failures (ffmpeg missing, probe 5xx,
  storage blips) never demote previously-validated media; concurrent
  revalidation is single-flight instead of racing into 500s.
- Reconciliation: HEAD-confirm before demotion, zero-listing abort guard,
  per-track fault isolation, tracks/ prefix listing, first-run duplicate
  detection, metadata-only Ready refresh, 202 + background admin runs with
  a process-wide run guard, configurable worker remediation.
- API leak fixes: TracksV1 visibility masking + proxied URLs, hot-this-week
  proxied URLs, catalog cache no longer mutated per-request, truthful
  download expiry.
- Telemetry: null-element and closed-set validation, PostHog sink wired
  with failure isolation, prod log-level override.
- Rate limiting: playback-specific session-aware policy (default 300/min),
  limiter moved after authentication, CORS preflight max-age.
- Config: render.yaml PlaybackMedia secret slots; startup requires distinct
  https-only keys; HttpClient request-URI logging capped at Warning.
- Migration 20260716201351 now also creates TrackLyrics.Version (was in the
  model snapshot but never migrated).
- Contracts regenerated (Swashbuckle, Testing env); drift gate passes: added
  ProducesResponseType intent attributes across albums, behind-the-track,
  proof-videos, collections, entitlements, keys, register, admin stubs, and
  fixed the drift script's typeof-paren regex bug.
- Tests: +133 (validation gates, probe signatures, enforcement + range
  semantics, reconciliation safety, readiness/publish gate, ticket forgery,
  telemetry rejection). Suite: 1400 passed / 0 failed (Docker suites
  pending a separate run).

Docs: docs/PLAYBACK_FLOW_MAP.md (cited flow map + prod probe evidence),
docs/PRODUCTION_AUDIO_PLAYBACK_CONTRACT.md (rollout gates).

+++ src/Cambrian.Application/DTOs/Catalog/PublicCatalogTrackDto.cs
@@ -75,8 +75,8 @@ public sealed class PublicCatalogTrackDto
 
     public string? Artist { get; init; }
 
-    /// <summary>Lifetime public play count (preview + full plays, incl. anonymous).</summary>
-    public int Plays { get; init; }
+    /// <summary>Lifetime qualified-play count.</summary>
+    public long Plays { get; init; }
 
     /// <summary>Number of completed (paid) purchases for this track.</summary>
     public int Sales { get; init; }
