+++ src/Cambrian.Application/DTOs/Catalog/TrackStats.cs
@@ -1,13 +1,13 @@
 namespace Cambrian.Application.DTOs.Catalog;
 
 /// <summary>
-/// Live engagement counts for a single track, aggregated from the StreamSessions
-/// and Purchases tables. Computed on read; never stored on the Track row.
+/// Live engagement counts for a single track. Qualified plays come from the
+/// transactionally maintained TrackStats projection; sales come from completed purchases.
 /// </summary>
 public sealed class TrackStats
 {
-    /// <summary>Number of recorded stream sessions (plays) for the track.</summary>
-    public int Plays { get; set; }
+    /// <summary>Number of qualified plays for the track.</summary>
+    public long Plays { get; set; }
 
     /// <summary>Number of completed (paid) purchases for the track.</summary>
     public int Sales { get; set; }
