+++ src/Cambrian.Application/DTOs/Creator/CreatorDashboardResponse.cs
@@ -4,7 +4,7 @@ public class CreatorDashboardResponse
 {
     public long EarningsCents { get; set; }
     public long WeeklyEarningsCents { get; set; }
-    public int TotalPlays { get; set; }
+    public long TotalPlays { get; set; }
     public int TotalSales { get; set; }
     public decimal ConversionRate { get; set; }
     public List<TrackStatsDto> Tracks { get; set; } = new();
@@ -42,7 +42,7 @@ public class TrackStatsDto
     public string Title { get; set; } = "";
     public string? CoverArtUrl { get; set; }
     public int Sales { get; set; }
-    public int Plays { get; set; }
+    public long Plays { get; set; }
     public long EarnedCents { get; set; }
 }
